% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carly_function.R
\name{carly_function}
\alias{carly_function}
\title{Give me a sentence with my favorite dog breed}
\usage{
carly_function(dog_breed)
}
\arguments{
\item{dog_breed}{insert your favorite dog breed}
}
\description{
Give me a sentence with my favorite dog breed
}
\examples{
"lab", "golden retriever"
}
