% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msy.R
\name{msy}
\alias{msy}
\title{Calculate Maximum Sustainable Yield}
\usage{
msy(k, r)
}
\arguments{
\item{k}{is the carrying capacity in individuals}

\item{r}{is the intrinsic growth rate per year}
}
\value{
Returns the maximum sustainable yield
}
\description{
Calculate Maximum Sustainable Yield
}
\examples{
msy(3000, 0.06)
}
